pyproject.toml
[project]
name = "pythonproject3"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=22.0",
    "python-dotenv>=1.0",
    "telethon>=1.36",
]
